hvc_sysfs: change hyp trace log size to uint32_t
The type of the hypevisor log size is changing to uint32_t.
Modifying the hypervisor control driver accordingly.

JIRA VFND-2811

Reviewed-on: http://git-master/r/1253012
Reviewed-on: http://git-master/r/1314846
GVS: Gerrit_Virtual_Submit
Tested-by: Aniruddha Banerjee <aniruddhab@nvidia.com>

+++ drivers/virt/tegra/hvc_sysfs.c
@@ -70,7 +70,7 @@ static int hvc_sysfs_log(struct kobject *kobj)
 	}
 
 	log_ipa = info->log_ipa;
-	log_attr.size = info->log_size;
+	log_attr.size = (size_t)info->log_size;
 
 	iounmap(info);
